Decode matrix & multiply vector on the fly
Includes a rudimentary benchmark that starts from a dummy vector and
then multiplies the same 1024x1024 matrix to it 10_000 times in
sequence. Each iteration creates the lookup table from scratch, decodes
decodes the matrix again, and multiplies it to the vector while decoding
it. The time for loading the compressed matrix onto the GPU is not
included in the benchmark.

The benchmark takes about 8 seconds, i.e., 0.8 ms per decode+multiply
cycle. Of this, the creation of the lookup table takes only about
0.01ms, the rest is for decoding + multiplying.