Fix integer precisions in WebGL
It turns out that the code in the previous commit relied on integer
precisions being higher than what's guaranteed by the standard. While
this is worked on both tested desktop GPUs and an iPad, it failed on a
(relatively old) android phone, which apparently uses tighter (but still
valid) integer precisions. This commit fixes the issue by emulating wide
integer operations with a sequence of operations on shorter integers
that are guaranteed to be available on all compliant WebGl 2
implementations.

This commit also switches from 12-bit to 16-bit probabilities, as the
resulting wide integer operations are easier to be emulated with the
available short integer representations.

This commit seems to have an impact on run time. Decoding 100 matrices
with dimensions 1024x1024 now takes about 150 ms. However, this
implementation is still just a proof of concept, and a more careful
implementation can probably speed up decoding quite a bit.