Fix encoder bug due to breaking change in numpy
Due to some change in numpy, some calculations in the python encoder
that were previously performed in big-int representation were now
silently performed in fixed-width signed integer precision, which lead
to errors in semi-edge cases. Changing to unsigned integers fixes this.

This commit also adds a documentation of the binary data format in the
notebook `mock-data.ipynb`.