Simplify file format
Makes the encoded representation of each matrix independent from the
other matrices by flushing the coder states for all matrices because
this might enable future decoder run-time optimizations, and it turns
out that the overhead in bitrate from flushing is negligible anyway.

Also, reduce precision of probabilities in the entropy model to 8 bit
based on measurements that this only leads to about 2% overhead (and
will make lookup tables on the decoder much smaller).