Fix CPU reference implementation of vec-mat-mul
It turns out the WebGPU implementation of compressed matrix-vector
multiplication in commit f9d5b05 was actually correct. The discrepancy
to the results I got on the CPU were just because the CPU and GPU
implementations used slightly different rounding operations. The CPU
implementation resolved ties (numbers `xxx.5`) away from zero, whereas
the GPU implementation resolves ties towards the even neighbor (which is
correct behavior according to the WGSL spec [1]). Since resolving ties
towards even is probably a slightly better choice anyway, this commit
leaves the GPU implementation as is and aligns the CPU implementation to
it.

Also adds debugging infrastructure to the binaries `verify-random` and
`demo` that allows writing out intermediate vectors and uncompressed
matrices to a `safetensors` file (using flags `--debug` and
`--debug_matrices`).

[1] https://www.w3.org/TR/WGSL/#round-builtin